refactor: remove plural
