+++ cmd/application/use-case/verify-notification/verify-notification.go
@@ -1,25 +1,25 @@
 package verifynotificationusecase
 
 import (
-	domaindto "go-clean-api/cmd/domain/dto"
-	portsservice "go-clean-api/cmd/domain/services"
-	domainusecases "go-clean-api/cmd/domain/use-case"
+	"go-clean-api/cmd/domain/dto"
+	"go-clean-api/cmd/domain/service"
+	usecase "go-clean-api/cmd/domain/use-case"
 	"log"
 )
 
 type (
 	notifyUseCase struct {
-		NotificationService portsservice.NotificationService
+		NotificationService service.NotificationService
 	}
 )
 
-func New(ns portsservice.NotificationService) domainusecases.NotifyUserUseCase {
+func New(ns service.NotificationService) usecase.NotifyUserUseCase {
 	return &notifyUseCase{
 		NotificationService: ns,
 	}
 }
 
-func (nuc *notifyUseCase) Notify(dto domaindto.Event) error {
+func (nuc *notifyUseCase) Notify(dto dto.Event) error {
 
 	log.Default().Println("amqp consumer completed with succeffully")
