+++ cmd/application/use-case/verify-notification/verify-notification_test.go
@@ -3,7 +3,7 @@ package verifynotificationusecase
 import (
 	domaindto "go-clean-api/cmd/domain/dto"
 	"go-clean-api/cmd/domain/exception"
-	mockservicesnotification "go-clean-api/cmd/shared/mocks/infra/services/notification"
+	mockservicesnotification "go-clean-api/cmd/shared/mocks/infra/service/notification"
 	"testing"
 )
