+++ cmd/domain/dto/register-user.dto.go
@@ -1,4 +1,4 @@
-package domaindto
+package dto
 
 type (
 	Dto struct {
