--- cmd/domain/services/notification.go
@@ -1,14 +1,0 @@
-package domainservice
-
-import (
-	domaindto "go-clean-api/cmd/domain/dto"
-)
-
-type (
-	NotificationService interface {
-		SendNotify(dto domaindto.Event) error
-		CheckNotify(msg string) error
-		SaveNotify(domaindto.Event) string
-		FindById(id string) (*domaindto.Event, error)
-	}
-)
